// frontend/src/types/generated/ImprovementsResponse.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { BaseImprovements } from "./BaseImprovements";
import type { MedalImprovements } from "./MedalImprovements";
import type { OperatorImprovements } from "./OperatorImprovements";
import type { RoguelikeThemeImprovement } from "./RoguelikeThemeImprovement";
import type { SandboxImprovements } from "./SandboxImprovements";
import type { StageImprovements } from "./StageImprovements";

export type ImprovementsResponse = { uid: string; stages: StageImprovements; roguelike: Array<RoguelikeThemeImprovement>; sandbox: SandboxImprovements; medals: MedalImprovements; operators: OperatorImprovements; base: BaseImprovements };