// frontend/src/types/generated/LeaderboardMover.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.

export type LeaderboardMover = { uid: string; nickname: string | null; nick_number: string | null; avatar_id: string | null; server: string; current_rank: number; previous_rank: number; rank_delta: number; current_score: number | null; score_delta: number | null };